requirements.txt
openai
langgraph
langchain-openai
langchain-core
fastapi
uvicorn[standard]
rich
tenacity
python-dotenv
aiosqlite
